+++ com.unity.cinemachine/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - InputAxisController has the option to suppress input while the attached camera is blending.
 - Added CinemachineCameraEvents and CinemachineBrainEvents behaviours for event processing.
 - Added BlendFinished and CameraDeactivated events.
+- Add split screen sample for the input system.
 - Samples UI works with both built-in Input System.
 - New 2D platformer sample showing custom camera manager.